[Refactor] Decouple MorselQueue from OLAP scan state
