+++ be/src/exec/CMakeLists.txt
@@ -243,6 +243,7 @@ set(PIPELINE_OPERATORS_BATCH_SRCS
     pipeline/scan/morsel.cpp
     pipeline/scan/morsel_queue.cpp
     pipeline/scan/morsel_queue_factory.cpp
+    pipeline/scan/olap_morsel_queue.cpp
     pipeline/scan/scan_morsel.cpp
     pipeline/scan/split_morsel_queue.cpp
     pipeline/scan/split_scan_morsel.cpp
