+++ be/src/exec/pipeline/scan/bucket_sequence_morsel_queue.cpp
@@ -16,6 +16,8 @@
 
 #include <fmt/compile.h>
 
+#include "gutil/casts.h"
+
 namespace starrocks::pipeline {
 
 BucketSequenceMorselQueue::BucketSequenceMorselQueue(MorselQueuePtr&& morsel_queue)
