+++ be/src/exec/pipeline/scan/connector_scan_operator.cpp
@@ -23,6 +23,7 @@
 #include "exec/pipeline/scan/balanced_chunk_buffer.h"
 #include "exec/pipeline/scan/scan_morsel.h"
 #include "exprs/expr_executor.h"
+#include "gutil/casts.h"
 #include "runtime/descriptors.h"
 #include "runtime/exec_env.h"
 #include "runtime/global_dict/parser.h"
