+++ be/src/exec/pipeline/scan/connector_scan_operator.h
@@ -19,6 +19,7 @@
 #include "exec/pipeline/scan/balanced_chunk_buffer.h"
 #include "exec/pipeline/scan/scan_operator.h"
 #include "exec/workgroup/work_group_fwd.h"
+#include "gutil/casts.h"
 #include "storage/chunk_helper.h"
 
 namespace starrocks {
